添加了时间设置功能

+++ System/MyRTC.c
@@ -1,7 +1,7 @@
 #include "stm32f10x.h"                  // Device header
 #include <time.h>
 
-uint16_t MyRTC_Time[] = {2025, 5, 27, 10, 23, 55};	//定义全局的时间数组，数组内容分别为年、月、日、时、分、秒
+int MyRTC_Time[] = {2025, 5, 27, 10, 23, 55};	//定义全局的时间数组，数组内容分别为年、月、日、时、分、秒
 
 void MyRTC_SetTime(void);				//函数声明
